+++ src/main/java/me/pooriya/plotfour/board/creator/SimpleBoardCreator.java
@@ -1,4 +1,4 @@
-package me.pooriya.plotfour.board;
+package me.pooriya.plotfour.board.creator;
 
 public class SimpleBoardCreator implements BoardCreator {
