+++ src/test/java/me/pooriya/plotfour/board/creator/SimpleBoardCreatorTest.java
@@ -1,4 +1,4 @@
-package me.pooriya.plotfour.board;
+package me.pooriya.plotfour.board.creator;
 
 import org.junit.Test;
